Refactor R2 notification handlers/options
