+++ packages/wrangler/src/r2/index.ts
@@ -9,27 +9,21 @@ import { FatalError, UserError } from "../errors";
 import { CommandLineArgsError, printWranglerBanner } from "../index";
 import { logger } from "../logger";
 import * as metrics from "../metrics";
-import { getQueueByID } from "../queues/client";
-import { requireApiToken, requireAuth } from "../user";
+import { requireAuth } from "../user";
 import { MAX_UPLOAD_SIZE } from "./constants";
 import {
-	actionsForEventCategories,
 	bucketAndKeyFromObjectPath,
 	createR2Bucket,
-	deleteEventNotificationConfig,
 	deleteR2Bucket,
 	deleteR2Object,
-	getEventNotificationConfig,
 	getR2Object,
 	listR2Buckets,
-	putEventNotificationConfig,
 	putR2Object,
-	tableFromNotificationGetResponse,
 	usingLocalBucket,
 } from "./helpers";
+import * as Notification from "./notification";
 import * as Sippy from "./sippy";
 import type { CommonYargsArgv } from "../yargs-types";
-import type { R2EventType } from "./helpers";
 import type { R2PutOptions } from "@cloudflare/workers-types/experimental";
 
 const CHUNK_SIZE = 1024;
@@ -555,130 +549,20 @@ export function r2(r2Yargs: CommonYargsArgv) {
 						.command(
 							"get <bucket>",
 							"Get event notification configuration for a bucket",
-							(yargs) => {
-								return yargs.positional("bucket", {
-									describe:
-										"The name of the bucket for which notifications will be emitted",
-									type: "string",
-									demandOption: true,
-								});
-							},
-							async (args) => {
-								await printWranglerBanner();
-								const config = readConfig(args.config, args);
-								const accountId = await requireAuth(config);
-								const apiCreds = requireApiToken();
-								const resp = await getEventNotificationConfig(
-									apiCreds,
-									accountId,
-									`${args.bucket}`
-								);
-								const tableOutput = await tableFromNotificationGetResponse(
-									config,
-									resp[args.bucket],
-									getQueueByID
-								);
-								logger.table(tableOutput);
-							}
+							Notification.GetOptions,
+							Notification.GetHandler
 						)
 						.command(
 							"create <bucket>",
 							"Create new event notification configuration for an R2 bucket",
-							(yargs) => {
-								return yargs
-									.positional("bucket", {
-										describe:
-											"The name of the bucket for which notifications will be emitted",
-										type: "string",
-										demandOption: true,
-									})
-									.option("event-types", {
-										describe:
-											"Specify the kinds of object events to emit notifications for. ex. '--event-types object-create object-delete'",
-										alias: "event-type",
-										choices: Object.keys(actionsForEventCategories),
-										demandOption: true,
-										requiresArg: true,
-										type: "array",
-									})
-									.option("prefix", {
-										describe:
-											"only actions on objects with this prefix will emit notifications",
-										requiresArg: false,
-										type: "string",
-									})
-									.option("suffix", {
-										describe:
-											"only actions on objects with this suffix will emit notifications",
-										type: "string",
-									})
-									.option("queue", {
-										describe:
-											"The name of the queue to which event notifications will be sent. ex '--queue my-queue'",
-										demandOption: true,
-										requiresArg: true,
-										type: "string",
-									});
-							},
-							async (args) => {
-								await printWranglerBanner();
-								const config = readConfig(args.config, args);
-								const accountId = await requireAuth(config);
-								const apiCreds = requireApiToken();
-								const {
-									bucket,
-									queue,
-									eventTypes,
-									prefix = "",
-									suffix = "",
-								} = args;
-								await putEventNotificationConfig(
-									config,
-									apiCreds,
-									accountId,
-									`${bucket}`,
-									`${queue}`,
-									eventTypes as R2EventType[],
-									`${prefix}`,
-									`${suffix}`
-								);
-								logger.log("Configuration created successfully!");
-							}
+							Notification.CreateOptions,
+							Notification.CreateHandler
 						)
 						.command(
 							"delete <bucket>",
 							"Delete event notification configuration for an R2 bucket and queue",
-							(yargs) => {
-								return yargs
-									.positional("bucket", {
-										describe:
-											"The name of the bucket for which notifications will be emitted",
-										type: "string",
-										demandOption: true,
-									})
-									.option("queue", {
-										describe:
-											"The name of the queue that is configured to receive notifications. ex '--queue my-queue'",
-										demandOption: true,
-										requiresArg: true,
-										type: "string",
-									});
-							},
-							async (args) => {
-								await printWranglerBanner();
-								const config = readConfig(args.config, args);
-								const accountId = await requireAuth(config);
-								const apiCreds = requireApiToken();
-								const { bucket, queue } = args;
-								await deleteEventNotificationConfig(
-									config,
-									apiCreds,
-									accountId,
-									`${bucket}`,
-									`${queue}`
-								);
-								logger.log("Configuration deleted successfully!");
-							}
+							Notification.DeleteOptions,
+							Notification.DeleteHandler
 						);
 				}
 			);
